+++ lib/src/renderer/droplets-renderer.cpp
@@ -93,11 +93,11 @@ namespace EdgeLighting
         return true;
     }
 
-    void DropletsRenderer::Update(float, float, const Config &)
+    void DropletsRenderer::Update(float, double, const Config &)
     {
     }
 
-    void DropletsRenderer::Render(int viewportWidth, int viewportHeight, float time, const Config &config)
+    void DropletsRenderer::Render(int viewportWidth, int viewportHeight, double time, const Config &config)
     {
         if (!config.droplets.enable || viewportWidth <= 0 || viewportHeight <= 0)
         {
@@ -134,7 +134,18 @@ namespace EdgeLighting
         mShaderProgram.SetUniform("uRectSize", glm::vec2(config.geometry.width, config.geometry.height));
         mShaderProgram.SetUniform("uRectCenter", rectCenter);
         mShaderProgram.SetUniform("uCornerRadius", GeometryUtils::GetEffectiveCornerRadius(config.geometry));
-        mShaderProgram.SetUniform("uTime", time);
+        // RAW, and deliberately not reduced. Unlike the neon and flare layers,
+        // this shader does not use time only as a phase: `uv.y += t * 0.75`
+        // feeds floor(uv * grid) into a per-cell hash, so removing any amount
+        // of time moves every drop to a different cell and reshuffles the whole
+        // field visibly. There is no period to wrap at.
+        //
+        // The consequence is recorded rather than hidden: the float uniform
+        // stops resolving a frame of drop motion at roughly 39 hours of
+        // uptime, and the rain slows and then stands still. Curing it means
+        // making the droplet field periodic in uv.y, which is a change to what
+        // the effect looks like. See review-findings I33b.
+        mShaderProgram.SetUniform("uTime", static_cast<float>(time));
         mShaderProgram.SetUniform("uAmount", config.droplets.amount);
         mShaderProgram.SetUniform("uSpeed", config.droplets.speed);
         mShaderProgram.SetUniform("uLanes", std::max(config.droplets.lanes, 1));
